fixing windows build
