+++ terminal_nosysioctl.go
@@ -1,3 +1,4 @@
+//go:build plan9 || solaris
 // +build plan9 solaris
 
 package goterm
@@ -9,4 +10,13 @@ func getWinsize() (*winsize, error) {
 	ws.Row = 24
 
 	return ws, nil
+}
+
+// Height gets console height
+func Height() int {
+	ws, err := getWinsize()
+	if err != nil {
+		return -1
+	}
+	return int(ws.Row)
 }
