+++ terminal_windows.go
@@ -1,8 +1,11 @@
+//go:build windows
 // +build windows
 
 package goterm
 
 import (
+	"errors"
+	"math"
 	"os"
 
 	"golang.org/x/sys/windows"
@@ -20,4 +23,18 @@ func getWinsize() (*winsize, error) {
 	ws.Row = uint16(info.Window.Bottom - info.Window.Top + 1)
 
 	return ws, nil
+}
+
+// Height gets console height
+func Height() int {
+	ws, err := getWinsize()
+	if err != nil {
+		// returns math.MinInt32 if we could not retrieve the height of console window,
+		// like VSCode debugging console
+		if errors.Is(err, windows.WSAEOPNOTSUPP) {
+			return math.MinInt32
+		}
+		return -1
+	}
+	return int(ws.Row)
 }
